terra: Refactor ExchangeRateOracle
- Factor types into their own modules under `terra`
- Move exchange rate calculation to `Denom`
- Use `tokio::sync::Mutex`
